Made the bottom margin on the posts bigger
not a single blog post will display when you first open the page. also added a pic that i used for the post, we dont need to keep it. Also added some fake updates and blog posts, im not sure if thats the kind of content we are looking for though